Adding comparison data with additional variables
This is a version of the tabular data with additional variables from Crosas et al, such as "crosassource", which was used during data collection to record where the journal data policy came from. The analysis doesn't use these variables, so it won't be published with the rest of the data files.